add litellm_request_id so we can group
